+++ CHANGELOG.md
@@ -9,6 +9,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Add support for the `DTSTAMP` field.
   If not provided, it is initialized to the current UTC DateTime when serializing.
 
+- Add support for the `ATTACH` field.
+  If not provided
+
 ## v1.1.2 - 2022-01-16
 
 - Handle calendars that wrap output such as Google Calendar